a little commit with some changes before a big update

+++ include/window.h
@@ -18,9 +18,9 @@ public:
 	void Display();
 
 
-	float GetWindowWidth();
+	int GetWindowWidth();
 
-	float GetWindowHeight();
+	int GetWindowHeight();
 
 	bool running = true;
